requirements.txt
python-telegram-bot==21.4
google-generativeai==0.7.2
faiss-cpu==1.8.0.post1
pypdf==5.0.1
python-docx==1.1.2
youtube_transcript_api==0.6.2
requests==2.32.3
tiktoken==0.7.0
